crud/requirements.txt
django==3.1
django-import-export==2.5.0
